% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampleRandomEffectPrediction.R
\docType{data}
\name{exampleRandomEffectPrediction}
\alias{exampleRandomEffectPrediction}
\title{An Example of Data for Predicting the Random Effects}
\format{
a data table
\describe{
\item{provinceID}{A string that stands for the province}
\item{plotID}{A string that stands for the plot}
\item{treeID}{A string that stands for the tree}
\item{species_code}{A 6-character code for the species (See getSpeciesList function)}
\item{dbh_cm}{Diameter at breast height (cm)}
\item{Htot_m}{Tree height (m)}
\item{h}{Section height (m)}
\item{d_cm}{Section diameter (cm)}
}
}
\usage{
exampleRandomEffectPrediction
}
\description{
An Example of Data for Predicting the Random Effects
}
\keyword{datasets}
