% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volTaper.R
\name{getSpeciesList}
\alias{getSpeciesList}
\title{Provide the Species List}
\usage{
getSpeciesList()
}
\description{
The values in the fiels \code{species_code} must be used to select
the appropriate species.
}
