% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parms_ObsH.R
\docType{data}
\name{Parms.ObsH}
\alias{Parms.ObsH}
\title{parameters for available H_tot}
\format{
a data table
\describe{
\item{species_code}{combination of the 4-letter genus code and 3-letter species code (NFI_CODE)}
\item{fixed_b}{fixed effect}
\item{stddev_prov_b}{sd of random effect prov}
\item{stddev_idPlot_b}{sd of random effect idplot}
\item{stddev_idtree_b}{sd of random effect idtree}
\item{stddev_Residual}{sd of residuals}
}
}
\usage{
Parms.ObsH
}
\description{
contain the parameter for obsH
}
\examples{
data(Parms.ObsH)
}
\keyword{datasets}
