% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volTaper.R
\name{predictRandomEffects}
\alias{predictRandomEffects}
\title{Predict the random effects}
\usage{
predictRandomEffects(data)
}
\arguments{
\item{data}{a data.frame object properly formatted}
}
\description{
Predict the random effects
}
