% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_list.R
\docType{data}
\name{Species.List}
\alias{Species.List}
\title{species list}
\format{
a data table
\describe{
\item{species_code}{combination of the 4-letter genus code and 3-letter species code (NFI_CODE)}
\item{ENGLISH_NAME}{english name}
\item{FRENCH_NAME}{french name}
\item{SCIENTIFIC_NAME}{scientific name}
}
}
\usage{
Species.List
}
\description{
species list
}
\keyword{datasets}
