% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parms_ObsH.R
\docType{data}
\name{tab4.ObsH}
\alias{tab4.ObsH}
\title{parameters for available H_tot}
\format{
a data table
\describe{
\item{Hchoice}{obs or est}
\item{name_eng}{english name}
\item{CANFI_code}{canfi code}
\item{fixed_b}{fixed effect}
\item{stddev_prov_b}{sd of random effect prov}
\item{stddev_idPlot_b}{sd of random effect idplot}
\item{stddev_idtree_b}{sd of random effect idtree}
\item{stddev_Residual}{sd of residuals}
}
}
\usage{
tab4.ObsH
}
\description{
contain the parameter for obsH
}
\examples{
data(tab4.ObsH)
}
\keyword{datasets}
