% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parms_EstH.R
\docType{data}
\name{tab6.EstH}
\alias{tab6.EstH}
\title{parameters for unavailable H_tot}
\format{
a data table
\describe{
\item{Hchoice}{obs or est}
\item{name_eng}{english name}
\item{CANFI_code}{canfi code}
\item{a1}{fixed effect a1}
\item{a2}{fixed effect a2}
\item{b}{fixed effect b}
\item{stddev_prov_a2}{sd of random effect prov of a2}
\item{stddev_idPlot_a2}{sd of random effect idplot of a2}
\item{stddev_prov_b}{sd of random effect prov of b}
\item{stddev_idPlot_b}{sd of random effect idplot of b}
\item{stddev_idtree_b}{sd of random effect idtree of b}
\item{sddev_Residual}{sd of residuals}
}
}
\usage{
tab6.EstH
}
\description{
parameters for unavailable H_tot
}
\keyword{datasets}
