% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volTaper.R
\name{volTaper}
\alias{volTaper}
\title{Wood volume calculation using taper models}
\usage{
volTaper(data, measuredH = TRUE, limit.dob = NULL, limit.height = NULL)
}
\arguments{
\item{data}{a data frame that includes at least 4 columns: canfi, treeID, dbh_cm, Htot_m}

\item{measuredH}{logic, TRUE if Htot is available .}

\item{limit.dob}{numeric, the limit of dob (cm) .}

\item{limit.height}{numeric, the limit of height (m) .}
}
\description{
Description
}
\details{
details
}
\examples{
sample_data <- data.table::data.table(
  canfi = c(101,102),
  treeID = c(1,2),
  dbh_cm = c(30,30),
  Htot_m = c(18,18)
)
V_ObsH.R <- volTaper(data = sample_data, measuredH = TRUE, limit.dob = 9, limit.height = NULL)
V_EstH.R <- volTaper(data = sample_data, measuredH = FALSE, limit.dob = 9, limit.height = NULL)
}
\references{
Ung, C.-H.; Guo, X.J.; Fortin, M. 2013. Canadian national taper models. For. Chron. 89(2):211-224
}
\author{
XiaoJing Guo \email{xiaojing.guo@NRCan-RNCan.gc.ca}
}
