% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volTaper.R
\name{predictRandomEffects}
\alias{predictRandomEffects}
\title{Predict the random effects}
\usage{
predictRandomEffects(data, produceEnhancedPred = F)
}
\arguments{
\item{data}{a data.frame object properly formatted}

\item{produceEnhancedPred}{a logical if true a data.frame with enhanced predictions is
added to the returned list}
}
\value{
a list with the blups
}
\description{
Predict the random effects
}
