% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volTaper.R
\name{volTaper}
\alias{volTaper}
\title{Wood volume calculation using taper models}
\usage{
volTaper(
  data,
  forceHeightEstimation = FALSE,
  limit.dob = NULL,
  limit.height = NULL
)
}
\arguments{
\item{data}{a data frame that includes at least 4 columns: species_code, treeID, dbh_cm, Htot_m}

\item{forceHeightEstimation}{logic, TRUE to force the height estimation even if Htot is available.}

\item{limit.dob}{numeric, the limit of dob (cm) .}

\item{limit.height}{numeric, the limit of height (m) .}
}
\description{
Wood volume calculation using taper models
}
\examples{
sample_data <- data.table::data.table(
  species_code = c("PICEMAR","PICEGLA"),
  treeID = c(1,2),
  dbh_cm = c(30,30),
  Htot_m = c(18,18)
)
V_ObsH.R <- volTaper(data = sample_data, limit.dob = 9, limit.height = NULL)
V_EstH.R <- volTaper(data = sample_data, forceHeightEstimation = TRUE, limit.dob = 9, limit.height = NULL)
}
\references{
Ung, C.-H.; Guo, X.J.; Fortin, M. 2013. Canadian national taper models. For. Chron. 89(2):211-224
}
\author{
Xiao Jing Guo \email{xiaojing.guo@NRCan-RNCan.gc.ca}
}
